Fix missing local variables in threaded file loader, fix possible race conditions and non-thread-safe behavior.
The thread load_file function needed access to the reloader object,
which was a local variable in the function that set up the thread.
However, I think there's a possibility of race conditions and non-thread
safe activity if we update shared objects from inside the thread.  I
changed the work in the thread to collect everything in objects on the
class (and turned the thread in a Thread subclass).  When the thread is
finished running, we pull the loaded mounts out of it.  I also deferred
the server hooks until then because I'm not sure the author of the hooks
might not expect them to be run in a thread.

I also believe there's a potential race condition that the threads might
not be done executing by the time we were moving on to creating the
Router() object.  So we now wait up to 5 seconds for the threads to
finish before proceeding to use the result.